requirements.txt
appdirs==1.4.4
social-auth-app-django==5.0.0
asgiref==3.4.1
certifi==2020.12.5
distlib==0.3.1
Django==3.2.6
filelock==3.0.12
numpy==1.21.0
pipenv==2020.11.15
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
virtualenv==20.4.3
virtualenv-clone==0.5.4
django-crispy-forms == 1.12.0
pillow==8.3.1
django-recaptcha == 2.0.6
django-phonenumber-field[phonenumberslite] == 3.4.1
